tmpdir: new utility
designed to be used by package recipes, use_tmpdir creates and pushes to
destinations a temp directory. It is cleaned automatically by pkg_runner, or
you can clean it yourself with clean_tmpdir.

If use_tmpdir is called before clean_tmpdir is called again, it returns 1 and
does not create a new temp directory.