Fix for odd npm ls depth/parseable conflict
With some versions of npm, the `--depth 0` and `--parseable` options are
mutually exclusive -- `--parseable` will ignore the `--depth` option. It
seems to be fixed in the current version of npm, but I don't want to
take the presence of that for granted.

Instead we're doing a `--depth 0` and extracting from that.

Also, tests.